cleaned up landing and coin page
