+++ src/components/CoinDescription/index.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { SquareStackIcon } from "../CoinBox2/CoinBox2.styles";
+import {
+  DescriptionWrapper,
+  CoinDescriptionDiv,
+} from "./CoinDescription.styles";
+
+export const CoinDescription = ({ coinDescription }) => {
+  return (
+    <DescriptionWrapper>
+      <SquareStackIcon />
+      <CoinDescriptionDiv>{coinDescription}</CoinDescriptionDiv>
+    </DescriptionWrapper>
+  );
+};
