+++ src/components/CoinUrl/CoinUrl.styles.jsx
@@ -1,0 +1,56 @@
+import styled from "styled-components";
+import React from "react";
+
+export const UrlDiv = styled.div`
+  width: 32%;
+  height: 65px;
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  background: ${(props) => props.theme.main};
+  border-radius: 20px;
+`;
+const UrlSvg = styled.svg`
+  width: 20px;
+  cursor: pointer;
+`;
+export const UrlIcon = () => {
+  return (
+    <UrlSvg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      stroke="currentColor"
+      className="w-6 h-6"
+    >
+      <path
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M13.19 8.688a4.5 4.5 0 011.242 7.244l-4.5 4.5a4.5 4.5 0 01-6.364-6.364l1.757-1.757m13.35-.622l1.757-1.757a4.5 4.5 0 00-6.364-6.364l-4.5 4.5a4.5 4.5 0 001.242 7.244"
+      />
+    </UrlSvg>
+  );
+};
+const CopySvg = styled.svg`
+  width: 20px;
+  cursor: pointer;
+`;
+export const CopyIcon = () => {
+  return (
+    <CopySvg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      strokeWidth={1.5}
+      stroke="currentColor"
+      className="w-6 h-6"
+    >
+      <path
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        d="M15.75 17.25v3.375c0 .621-.504 1.125-1.125 1.125h-9.75a1.125 1.125 0 01-1.125-1.125V7.875c0-.621.504-1.125 1.125-1.125H6.75a9.06 9.06 0 011.5.124m7.5 10.376h3.375c.621 0 1.125-.504 1.125-1.125V11.25c0-4.46-3.243-8.161-7.5-8.876a9.06 9.06 0 00-1.5-.124H9.375c-.621 0-1.125.504-1.125 1.125v3.5m7.5 10.375H9.375a1.125 1.125 0 01-1.125-1.125v-9.25m12 6.625v-1.875a3.375 3.375 0 00-3.375-3.375h-1.5a1.125 1.125 0 01-1.125-1.125v-1.5a3.375 3.375 0 00-3.375-3.375H9.75"
+      />
+    </CopySvg>
+  );
+};
