+++ src/components/CoinUrl/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { UrlDiv, UrlIcon, CopyIcon } from "./CoinUrl.styles";
+
+export const CoinUrl = ({ blockchainSite }) => {
+  return (
+    <UrlDiv>
+      <UrlIcon />
+      <div>{blockchainSite}</div>
+      <CopyIcon />
+    </UrlDiv>
+  );
+};
