+++ src/components/PageHeader/PageHeader.styles.jsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const HeaderDiv = styled.div`
+  margin-bottom: 30px;
+  width: 98%;
+`;
