+++ src/components/TopChartTitlePrice/TopChartTitlePrice.styles.jsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const TopChartHeader = styled.div`
+  width: 50%;
+  padding-left: 55px;
+`;
