+++ src/pages/Coins/Coin/index.jsx
@@ -1,40 +1,19 @@
 import axios from "axios";
-import { CoinsPercentageBar } from "../../../components/CoinsPercentageBar";
-import { CurrencyConverter } from "../../../components/CurrencyConverter";
 import { useEffect, useContext, useReducer } from "react";
+import { CurrencyConverter } from "../../../components/CurrencyConverter";
 import { CurrencyContext } from "../../../contexts/CurrencyContext";
-import { formatNumber, formatDate, getRandomColor } from "../../../utilities";
-import { ArrowUp, ArrowDown } from "../Coins.styles";
+import { formatDate, getRandomColor, coinPricesData } from "../../../utilities";
+import { CoinBox1 } from "../../../components/CoinBox1";
+import { CoinBox2 } from "../../../components/CoinBox2";
+import { CoinBox3 } from "../../../components/CoinBox3";
+import { CoinDescription } from "../../../components/CoinDescription";
+import { CoinUrl } from "../../../components/CoinUrl";
+import { PageHeader } from "../../../components/PageHeader";
 import { BigLineChart } from "../../../components/LineChart";
 import {
   CoinPageWrapper,
-  HeaderDiv,
   SummaryWrapper,
-  CoinBoxContainer,
-  CoinBox,
-  CoinBoxInner,
-  CoinImage,
-  CoinUrlBox,
-  PriceBoxContainer,
-  CoinDataBoxContainer,
-  UrlIcon,
-  HomePageDiv,
-  PriceCHangePercentageDiv,
-  PriceChangePercentage,
-  SquareStackIcon,
-  AllTimeContainer,
-  AllTimeDiv,
-  ArrowSvgDiv,
-  AllTimeP,
-  CoinDataRow,
-  SmallRectangleBlue,
-  PlusSmallSvg,
-  DescriptionWrapper,
-  CoinDescriptionDiv,
   CoinUrlsRow,
-  UrlDiv,
-  CopyIcon,
-  UrlAddressDiv,
   CurrencyConversionRow,
 } from "./Coin.styles";
 
@@ -43,7 +22,6 @@ const ACTIONS = {
 };
 export const Coin = (props) => {
   const { currency } = useContext(CurrencyContext);
-
   const getCoinData = async (coinId, days, interval) => {
     try {
       const { data } = await axios(
@@ -60,7 +38,6 @@ export const Coin = (props) => {
       console.log(err);
     }
   };
-
   const reducer = (state, action) => {
     switch (action.type) {
       case ACTIONS.GET_COIN_DATA:
@@ -129,10 +106,6 @@ export const Coin = (props) => {
       coinPrices: [],
     },
   });
-  const getThemeColors = () => {
-    const theme = localStorage.getItem("theme");
-    return JSON.parse(theme);
-  };
   useEffect(() => {
     getCoinData(
       props.match.params.coinId,
@@ -140,214 +113,56 @@ export const Coin = (props) => {
       state.coinData.interval
     );
   }, [currency]);
-
   localStorage.setItem("coinCurrentPrice", state.coinData.coinCurrentPrice);
-  const theme = getThemeColors();
-  const coinPricesData = {
-    labels: state.coinData.chartLabels,
-    datasets: [
-      {
-        label: `${props.match.params.coinId} Price`,
-        data: state.coinData.coinPrices,
-        borderColor:
-          state.coinData.coinPrices[0] <
-          state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
-            ? theme.btcPriceChartBorderColorGain
-            : state.coinData.coinPrices[0] >
-              state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
-            ? theme.btcPriceChartBorderColorLoss
-            : theme.btcPriceChartBorderColorGain,
-        backgroundColor: (context) => {
-          const ctx = context.chart.ctx;
-          const gradient = ctx.createLinearGradient(0, 0, 0, 350);
-          state.coinData.coinPrices[0] <
-          state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
-            ? gradient.addColorStop(0, theme.btcPriceChartGradienColorGain)
-            : state.coinData.coinPrices[0] >
-              state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
-            ? gradient.addColorStop(0, theme.btcPriceChartGradienColorLoss)
-            : gradient.addColorStop(0, theme.btcPriceChartGradienColorGain);
-          gradient.addColorStop(1, "rgba(0, 0, 0, 0.0)");
-          return gradient;
-        },
-        pointRadius: 0,
-        borderWidth: 3,
-        fill: true,
-      },
-    ],
-  };
   return (
     <CoinPageWrapper>
-      <HeaderDiv>
-        <h2>Summary</h2>
-      </HeaderDiv>
+      <PageHeader text={"Summary"} />
       <SummaryWrapper>
-        <CoinBoxContainer>
-          <CoinBox>
-            <CoinBoxInner>
-              <CoinImage src={state.coinData.coinImgSrc} />
-            </CoinBoxInner>
-            <h2>
-              {state.coinData.coinName} ({state.coinData.coinSymbol})
-            </h2>
-          </CoinBox>
-          <CoinUrlBox>
-            <UrlIcon />
-            <HomePageDiv>{state.coinData.coinHomePage}</HomePageDiv>
-          </CoinUrlBox>
-        </CoinBoxContainer>
-        <PriceBoxContainer>
-          <h2 style={{ fontSize: "44px" }}>
-            {formatNumber(state.coinData.coinCurrentPrice)}
-          </h2>
-          <PriceCHangePercentageDiv>
-            <PriceChangePercentage
-              color={
-                state.coinData.priceChangePercentage24h > 0
-                  ? "#00fc2a"
-                  : "#fe1040"
-              }
-            >
-              {" "}
-              {state.coinData.priceChangePercentage24h}%
-            </PriceChangePercentage>
-            {state.coinData.priceChangePercentage24h > 0 ? (
-              <ArrowUp />
-            ) : (
-              <ArrowDown />
-            )}
-          </PriceCHangePercentageDiv>
-          <SquareStackIcon />
-          <AllTimeContainer>
-            <ArrowSvgDiv>
-              <ArrowUp />
-            </ArrowSvgDiv>
-            <AllTimeDiv>
-              <AllTimeP>
-                All Time High:{" "}
-                <span> {formatNumber(state.coinData.coinAth)}</span>
-              </AllTimeP>
-              <p>{state.coinData.coinAthDate}</p>
-            </AllTimeDiv>
-          </AllTimeContainer>
-          <AllTimeContainer>
-            <ArrowSvgDiv>
-              <ArrowDown />
-            </ArrowSvgDiv>
-            <AllTimeDiv>
-              <AllTimeP>
-                All Time Low:{" "}
-                <span> {formatNumber(state.coinData.coinAtl)}</span>
-              </AllTimeP>
-              <p>{state.coinData.coinAtlDate}</p>
-            </AllTimeDiv>
-          </AllTimeContainer>
-        </PriceBoxContainer>
-        <CoinDataBoxContainer>
-          <CoinDataRow>
-            <SmallRectangleBlue>
-              <PlusSmallSvg />
-            </SmallRectangleBlue>
-            <p>
-              Market Cap:{" "}
-              <span>{formatNumber(state.coinData.coinMarketCap)}</span>{" "}
-            </p>
-          </CoinDataRow>
-          <CoinDataRow>
-            <SmallRectangleBlue>
-              <PlusSmallSvg />
-            </SmallRectangleBlue>
-            <p>
-              Fully Diluted Valuation:{" "}
-              <span>
-                {formatNumber(state.coinData.coinFullyDillutedValuation)}
-              </span>
-            </p>
-          </CoinDataRow>
-          <CoinDataRow>
-            <SmallRectangleBlue>
-              <PlusSmallSvg />
-            </SmallRectangleBlue>
-            <p>
-              Trading Volume 24h:{" "}
-              <span>{formatNumber(state.coinData.coinVolume24h)}</span>
-            </p>
-          </CoinDataRow>
-          <CoinDataRow>
-            <SmallRectangleBlue>
-              <PlusSmallSvg />
-            </SmallRectangleBlue>
-            <p>
-              Volume / Market:{" "}
-              <span>{state.coinData.coinVolumeOverMarketCap}%</span>
-            </p>
-          </CoinDataRow>
-          <CoinDataRow>
-            <SmallRectangleBlue>
-              <PlusSmallSvg />
-            </SmallRectangleBlue>
-            <p>
-              Circulating Supply:{" "}
-              <span>
-                {state.coinData.coinCirculatingSupply}{" "}
-                {state.coinData.coinSymbol}
-              </span>
-            </p>
-          </CoinDataRow>
-          <CoinDataRow>
-            <SmallRectangleBlue>
-              <PlusSmallSvg />
-            </SmallRectangleBlue>
-            <p>
-              Max Supply:{" "}
-              <span>
-                {state.coinData.coinTotalSupply} {state.coinData.coinSymbol}
-              </span>
-            </p>
-          </CoinDataRow>
-          <CoinDataRow style={{ width: "50%", marginLeft: "-180px" }}>
-            <CoinsPercentageBar
-              num1={state.coinData.num1 + "%"}
-              num2={state.coinData.num2 + "%"}
-              width={state.coinData.num1 + "%"}
-              color1={state.coinData.color1}
-              color2={state.coinData.color2}
-              background1={state.coinData.color1}
-              background2={state.coinData.color2}
-            />
-          </CoinDataRow>
-        </CoinDataBoxContainer>
+        <CoinBox1
+          src={state.coinData.coinImgSrc}
+          coinName={state.coinData.coinName}
+          coinSymbol={state.coinData.coinSymbol}
+          coinHomePage={state.coinData.coinHomePage}
+        />
+        <CoinBox2
+          coinPrice={state.coinData.coinCurrentPrice}
+          priceChangePercentage={state.coinData.priceChangePercentage24h}
+          ath={state.coinData.coinAth}
+          athDate={state.coinData.coinAthDate}
+          atl={state.coinData.coinAtl}
+          atlDate={state.coinData.coinAtlDate}
+        />
+        <CoinBox3
+          coinMarketCap={state.coinData.coinMarketCap}
+          coinFullyDillutedValuation={state.coinData.coinFullyDillutedValuation}
+          coinVolume24h={state.coinData.coinVolume24h}
+          coinVolumeOverMarketCap={state.coinData.coinVolumeOverMarketCap}
+          coinCirculatingSupply={state.coinData.coinCirculatingSupply}
+          coinSymbol={state.coinData.coinSymbol}
+          coinTotalSupply={state.coinData.coinTotalSupply}
+          num1={state.coinData.num1 + "%"}
+          num2={state.coinData.num2 + "%"}
+          color1={state.coinData.color1}
+          color2={state.coinData.color2}
+        />
       </SummaryWrapper>
-      <HeaderDiv>
-        <h2>Description</h2>
-      </HeaderDiv>
-      <DescriptionWrapper>
-        <SquareStackIcon />
-        <CoinDescriptionDiv>
-          {state.coinData.coinDescription}
-        </CoinDescriptionDiv>
-      </DescriptionWrapper>
+      <PageHeader text={"Description"} />
+      <CoinDescription coinDescription={state.coinData.coinDescription} />
       <CoinUrlsRow>
-        <UrlDiv>
-          <UrlIcon />
-          <UrlAddressDiv>{state.coinData.coinBlockChainSite1}</UrlAddressDiv>
-          <CopyIcon />
-        </UrlDiv>
-        <UrlDiv>
-          <UrlIcon />
-          <UrlAddressDiv>{state.coinData.coinBlockChainSite2}</UrlAddressDiv>
-          <CopyIcon />
-        </UrlDiv>
-        <UrlDiv>
-          <UrlIcon />
-          <UrlAddressDiv>{state.coinData.coinBlockChainSite3}</UrlAddressDiv>
-          <CopyIcon />
-        </UrlDiv>
+        <CoinUrl blockchainSite={state.coinData.coinBlockChainSite1} />
+        <CoinUrl blockchainSite={state.coinData.coinBlockChainSite2} />
+        <CoinUrl blockchainSite={state.coinData.coinBlockChainSite3} />
       </CoinUrlsRow>
       <CurrencyConversionRow>
         <CurrencyConverter coinSymbol={state.coinData.coinSymbol} />
       </CurrencyConversionRow>
-      <BigLineChart data={coinPricesData} />
+      <BigLineChart
+        data={coinPricesData(
+          state.coinData.chartLabels,
+          props.match.params.coinId,
+          state.coinData.coinPrices
+        )}
+      />
     </CoinPageWrapper>
   );
 };
